+++ spec/activerecord-multi-tenant/model_extensions_spec.rb
@@ -160,4 +160,17 @@ describe MultiTenant do
       end
     end
   end
+
+  describe '.with_lock' do
+    it 'supports with_lock blocks inside the block' do
+      @account = Account.create!(name: 'foo')
+
+      MultiTenant.with @account do
+        project = @account.projects.create!(name: 'project')
+        project.with_lock do
+          expect(project.name).to eq 'project'
+        end
+      end
+    end
+  end
 end